+++ Sources/PippinCore/Components/DebugMenuPresenter.swift
@@ -9,5 +9,7 @@ import Foundation
 
 /// Protocol providing a single method for a conforming debugging utility to inject its UI at the time of calling.
 public protocol DebugMenuPresenter: EnvironmentallyConscious {
-    func installViews()
+    /// Installs debugging views and controls for any present Pippin components that supply them.
+    /// - Parameter appControlPanel: optional views supplied by the app for debugging its business logic.
+    func installViews(appControlPanel: UIView?)
 }