+++ Sources/PippinDebugging/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Added
+
+- Ability to pass a custom view for app-specific debug views when installing Pippin debugging views for the menu.
+
 ## [2.0.0] 2020-03-10
 
 ### Changed